Lesson types moved into datatypes
