+++ client/tsconfig.json
@@ -11,7 +11,8 @@
         "sourceMap": true,
         "jsx": "react",
         "lib": ["dom", "es6", "es2015.promise"],
-        "types": ["whatwg-fetch", "material-ui", "history"]
+        "types": ["whatwg-fetch", "material-ui", "history"],
+        "baseUrl": "src"
     },
     "include": [
         "src/**/*.{tsx,ts}"
